str_char str_rchar
ints to longs on bin and str

tests updated


git-svn-id: file:///source/coreland/corelib/trunk@673 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ bin.h
@@ -1,8 +1,8 @@
 #ifndef BIN_H
 #define BIN_H
 
-int bin_chr(const char *, unsigned long, int);
-int bin_rchr(const char *, unsigned long, int);
+long bin_chr(const char *, unsigned long, int);
+long bin_rchr(const char *, unsigned long, int);
 
 int bin_char(const char *, unsigned long, int, unsigned long *);
 int bin_rchar(const char *, unsigned long, int, unsigned long *);
